+++ packages/blocks/core/CHANGELOG.md
@@ -123,6 +123,7 @@
 - `Social Links` block: support improved.
 - `Spacer` block: support improved.
 - `Table` block: support improved.
+- `Tag Cloud` block: support improved.
 
 
 ### Automated Tests
@@ -183,6 +184,7 @@
 - Updated E2E tests to check support of `Social Links` block.
 - Updated E2E tests to check support of `Spacer` block.
 - Updated E2E tests to check support of `Table` block.
+- Updated E2E tests to check support of `Tag Cloud` block.
 
 
 ## 1.3.0 (2025-04-27)